clean up and removed tooltip
