+++ src/features/dev/index.ts
@@ -1,1 +1,2 @@
 export * from './hooks';
+export { default as CmColorPicker } from './CmColorPicker';
